Migrate putnam (rotated-SOVC PDF, two trailing layouts)
